More tweaking; also updated the pubspec files as per beta2.
