+++ ng2/2-blankbadge/lib/components/pirate_badge.dart
@@ -6,11 +6,7 @@ import 'package:angular2/angular2.dart';
 
 @Component(
     selector: 'pirate-badge',
-    templateUrl: 'pirate_badge_component.html')
+    templateUrl: 'pirate_badge.html')
 class PirateBadge {
   String badgeName = "";
-
-  void updateBadge(String inputName) {
-    badgeName = inputName.toString();
-  }
 }
