+++ ng2/2-blankbadge/lib/components/pirate_badge.html
