+++ ng2/3-inputbadgename/lib/components/pirate_badge.dart
@@ -6,7 +6,11 @@ import 'package:angular2/angular2.dart';
 
 @Component(
     selector: 'pirate-badge',
-    templateUrl: 'pirate_badge_component.html')
+    templateUrl: 'pirate_badge.html')
 class PirateBadge {
   String badgeName = "";
+
+  void updateBadge(String inputName) {
+    badgeName = inputName.toString();
+  }
 }
