+++ ng2/3-inputbadgename/lib/components/pirate_badge.html
