+++ ng2/4-buttonbadge/lib/components/pirate_badge.html
