+++ ng2/4-buttonbadge/web/index.html
@@ -10,16 +10,11 @@
 
 <html>
   <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
     <title>Avast, Ye Pirates</title>
 
     <link rel="stylesheet" href="styles.css">
 
-    <script defer type="application/dart" src="main.dart"></script>
+    <script defer src="main.dart" type="application/dart"></script>
     <script defer src="packages/browser/dart.js"></script>
   </head>
   <body>
