+++ ng2/4-buttonbadge/web/styles.css
@@ -58,4 +58,3 @@ input[type="text"] {
         margin-left: 0;
     }
 }
-
