+++ ng2/5-piratenameservice/lib/components/pirate_badge.dart
@@ -2,32 +2,19 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:async';
-import 'package:angular2/angular2.dart' show Component, OnInit;
+import 'package:angular2/angular2.dart';
 import '../services/pirate_name.dart' show PirateNameService;
 
 @Component(
     selector: 'pirate-badge',
-    templateUrl: 'pirate_badge_component.html')
-class PirateBadge implements OnInit {
+    templateUrl: 'pirate_badge.html')
+class PirateBadge {
   String badgeName = "";
   String buttonText = "Aye! Gimme a name!";
-  bool enableButton = false;
-  bool enableInput = false;
+  bool enableButton = true;
+  bool enableInput = true;
 
-  ngOnInit() async {
-    try {
-      await PirateNameService.readyThePirates();
-      //on success
-      enableButton = true;
-      enableInput = true;
-    } catch (arrr) {
-      badgeName = 'Arrr! No names.';
-      print('Error initializing pirate names: $arrr');
-    }
-  }
-
-  generateBadge() => setBadgeName(new PirateNameService());
+  void generateBadge() => setBadgeName(new PirateNameService());
 
   void setBadgeName(PirateNameService newName) {
     if (newName == null) return;
