+++ ng2/5-piratenameservice/lib/components/pirate_badge.html
