+++ ng2/5-piratenameservice/pubspec.yaml
@@ -1,9 +1,15 @@
 name: pirate_badge
 description: Write a Dart web app code lab
 version: 0.0.1
+environment:
+  sdk: '>=1.13.0 <2.0.0'
 dependencies:
-  angular2: ^2.0.0-beta.0
+  angular2: 2.0.0-beta.2
   browser: ^0.10.0
 transformers:
 - angular2:
+    platform_directives:
+    - 'package:angular2/common.dart#COMMON_DIRECTIVES'
+    platform_pipes:
+    - 'package:angular2/common.dart#COMMON_PIPES'
     entry_points: web/main.dart
