+++ ng2/6-readjsonfile/lib/components/pirate_badge.html
